Attempted challenge 1 and 2. in challenge 1, i used the for loop to iterate through the array and passed  10 positions (0 to 9) to the countEvenNumber variable. An if statement was used to check if the value in the array is not a number, if the condition is true, the element is skipped.
in challenge 2, I also used the  for loop to iterate through the array.
An if statement was used to check if the value in the array is not a number, if true, the element is skipped.
and an else if condition was used to check if the value of element in array is not an even number, if true, the element is skipped.